discord-token-onliner-main/requirements.txt
discord.py-self
colorama